Small bug with the previous hex value before a replace call being logged as a char not hex.

+++ hexedit.cpp
@@ -123,7 +123,7 @@ void replace(Case current_case, string filename){
   char c = infile.peek();
   stringstream ss;
   //Prepare printout
-  ss << "Replacing char at pos " << posString << ": " << hex << c << " --> " << replacement;
+  ss << "Replacing char at pos " << posString << ": " << hex << (int)c << " --> " << replacement;
 
   //Put the character at that position
   infile.put(h);
